Add docker file to the repo
This is, I hope, the right approach.  The docker docs give a lot of examples
and none of them seem to show what the right step is.  So, I'm assuming that
this will work.  It is a bit of a chicken an egg and this allows me to make
sure the chicken defines what should be in the egg, so that when the image is
built it has what is needed.