Hard code things a bit for move
I moved from one laptop to another laptop, and for some quirky reason on the
new laptop the 'localName' of the device is no longer being picked up in the
advertisement data.  Not sure why.  So, since I only one one of these devices,
I opted to hard code it all back in.